move some functions to FAIR\Packages
